renamed stage directories
