+++ src/onemod/stage/data_stages/__init__.py
@@ -1,0 +1,3 @@
+from onemod.stage.data_stages.preprocessing_stage import PreprocessingStage
+
+__all__ = [PreprocessingStage]
