+++ src/onemod/stage/model_stages/spxmod_stage.py
